// hospital-application-ui/src/app/services/models/user.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Appointment } from '../models/appointment';
import { Feedback } from '../models/feedback';
import { GrantedAuthority } from '../models/granted-authority';
import { MedicalHistory } from '../models/medical-history';
import { Role } from '../models/role';
export interface User {
  accountLocked?: boolean;
  accountNonExpired?: boolean;
  accountNonLocked?: boolean;
  address?: string;
  authorities?: Array<GrantedAuthority>;
  cnp?: string;
  createdDate?: string;
  credentialsNonExpired?: boolean;
  dateOfBirth?: string;
  doctorAppointments?: Array<Appointment>;
  email?: string;
  enabled?: boolean;
  feedbackGiven?: Array<Feedback>;
  firstname?: string;
  fullName?: string;
  hospital?: string;
  id?: number;
  lastModifiedDate?: string;
  lastname?: string;
  medicalHistories?: Array<MedicalHistory>;
  name?: string;
  password?: string;
  patientAppointments?: Array<Appointment>;
  phoneNumber?: string;
  profilePicture?: string;
  roles?: Array<Role>;
  specialization?: string;
  username?: string;
}