// hospital-application-ui/src/app/services/models/admin-user-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AdminUserRequest {
  accountLocked?: boolean;
  address?: string;
  cnp?: string;
  dateOfBirth?: string;
  email?: string;
  enabled?: boolean;
  firstname?: string;
  hospital?: string;
  lastname?: string;
  phoneNumber?: string;
  role?: string;
  specialization?: string;
}